requirement.txt
streamlit==1.12.0
numpy==1.21.2
torch==1.10.0
Pillow==8.3.2
requests==2.26.0
opencv-python-headless==4.5.3.56
matplotlib==3.4.3
scipy==1.7.1
PyYAML==5.4.1
tqdm==4.62.3
seaborn==0.11.2
pandas==1.3.3
ultralytics==8.0.20
